requirements.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.2.4
django-crispy-forms==1.12.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-datatables==0.6.0
Pillow==8.2.0
psycopg2-binary==2.9.1
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
